requirements.txt
fastapi>=0.110.0,<0.112.0 
uvicorn[standard]>=0.27.0,<0.30.0
httpx>=0.26.0,<0.28.0
apscheduler[sqlalchemy]>=3.10.0,<4.0.0 
cryptography>=41.0.0,<44.0.0
python-dotenv>=1.0.0,<2.0.0
sqlalchemy>=1.4.0,<2.1.0 
aiosqlite>=0.19.0,<0.21.0 
requests>=2.30.0 
pytz 